refactor: extract daemon cli module
